# swaggerOut.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wahlhalla: Task Tracking API",
    "version": "0.0.1"
  },
  "swagger": "2.0",
  "paths": {
...
    "/api/subtask": {
      "get": {
        "tags": [
          "Subtask"
        ],
        "summary": "returns a list of all subtasks",
        "description": "Retrieves a list of all subtasks.",
        "responses": {
          "200": {
            "description": "a list of all subtasks ordered by last name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subtask"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/subtask/{id}": {
      "get": {
        "tags": [
          "Subtask"
        ],
        "summary": "returns a subtask by ID.",
        "description": "Retrieves subtask by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the subtask to retrieve.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a subtask found by specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subtask"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot find subtask with id={id}."
          },
          "500": {
            "description": "unknown error"
          }
        }
      },
      "put": {
        "tags": [
          "Subtask"
        ],
        "summary": "updates a subtask by ID.",
        "description": "Updates subtask by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the subtask to update.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Updated Subtask",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subtask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subtask was updated successfully."
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot find subtask with id={id}."
          },
          "500": {
            "description": "unknown error"
          }
        }
      },
      "delete": {
        "tags": [
          "Subtask"
        ],
        "summary": "deletes an subtask by the provided ID.",
        "description": "Deletes an subtask by the provided ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Numeric ID of the subtask to delete.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subtask was deleted successfully!"
          },
          "400": {
            "description": "validation error"
          },
          "404": {
            "description": "Cannot delete Subtask with id={id}. Maybe Subtask was not found!"
          },
          "500": {
            "description": "Could not delete subtask with id={id}"
          }
        }
      }
    },
    "/api/subtask/": {
      "post": {
        "tags": [
          "Subtask"
        ],
        "summary": "creates an subtask with the provided parameters.",
        "description": "Creates an subtask with the provided parameters",
        "requestBody": {
          "description": "Subtask To Create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subtask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the newly created subtask.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subtask"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/subtask/assignedTask/{assignedTaskId}": {
      "get": {
        "tags": [
          "Subtask"
        ],
        "summary": "returns subtasks by assigned task.",
        "description": "Retrieves subtasks by assigned task.",
        "parameters": [
          {
            "in": "path",
            "name": "assignedTaskId",
            "required": true,
            "description": "ID of the assigned task of the subtasks to retrieve.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "subtasks found by specified assigned task.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subtask"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/subtask/description/{description}": {
      "get": {
        "tags": [
          "Subtask"
        ],
        "summary": "returns subtasks by name.",
        "description": "Retrieves subtasks by name.",
        "parameters": [
          {
            "in": "path",
            "name": "description",
            "required": true,
            "description": "Description of the subtasks to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "subtasks found by specified description.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subtask"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
    "/api/subtask/assignedTask/{assignedTaskId}/description/{description}": {
      "get": {
        "tags": [
          "Subtask"
        ],
        "summary": "returns subtasks by assigned task and description.",
        "description": "Retrieves subtasks by assigned task and description.",
        "parameters": [
          {
            "in": "path",
            "name": "assignedTaskId",
            "required": true,
            "description": "ID of the assigned task of the subtasks to retrieve.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "description",
            "required": true,
            "description": "Description of the subtasks to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "subtasks found by specified assigned task and description.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subtask"
                  }
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "500": {
            "description": "unknown error"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}